Disable module 11.1.1.1.1
The test AsyncHttpToHttpViaActiveMqQueueTest fails intermittently when
multiple tests run at the same time in a single instance. When this
happens, it can be seen that eventhough the activeMQ urls are correctly
passed in the relevant deployment.properties files, CApps are created
using the ActiveMQ urls that belong to other test runs. The reson is
probably that the the default maven respository location(.m2) is being
shared by all the tests that run in a single system.
Relates to: https://github.com/wso2/product-ei/issues/3254