refactor(cli): update listing of artifacts
Refactor 'show' subcommands of all commands (related to MI artifacts) so that the printing logic is decoupled from the command's logic. In the existing implementation, code for printing tables is dependent on every command. With this change, an update to the table output can be introduced in one place such that it will be reflected across all commands.