Add MariaDB in GetDatabaseType
adding else if for MariaDB with type = "mysql"
when using the MariaDB JDBC driver we get an error:
Unsupported database: MariaDB. Database will not be created automatically by the WSO2 Registry. Please create the database using appropriate database scripts for the database.

With the MySQL JDBC and a MariaDB database it works fine.