Fix Service Catalog not publishing SOAP Proxy
The specific problem arises when the WSDLReader component in APIM
attempts to read the WSDL content with a schemaLocation attribute
set to 'proxy2?xsd=sample.xsd'. Upon parsing, WSDLReader treats this schemaLocation
as a file path rather than as a query string, leading to a failure in resolving the actual schema location.
So changed the schemaLocation into actual URL.

Fixes: https://github.com/wso2/micro-integrator/issues/2967