Fix registry not honouring cache expiration
The registry entries after cache expiration haven't been updated in the
run time. The issue was even though the expiration is detected the
version of the new registry entry is returned as same as old one then we
have returned the old value instead of changed new value. This is
because we have returned the saved time of file creation rather than
actual file modified time as the version.

Fixes: https://github.com/wso2/micro-integrator/issues/3083